requirements.txt
streamlit~=1.8.1
numpy~=1.22.3
tensorflow~=2.9.1
Pillow~=9.1.0